+++ lib/duckdb/data_chunk.rb
@@ -94,6 +94,9 @@ module DuckDB
         vector.assign_string_element(row_idx, value.to_s)
       when :blob
         vector.assign_string_element_len(row_idx, value.to_s)
+      when :timestamp
+        data = vector.get_data
+        MemoryHelper.write_timestamp(data, row_idx, value)
       else
         raise ArgumentError, "Unsupported type for DataChunk#set_value: #{type_id}"
       end